Adding solution for Day 9,2020 Parts One and Two
Part Two has an ugly brute force method for finding the
start and end of the contiguous sum matching the target
number. There are other ways, which I am thinking are more
efficient, which I may look at in the future. Also need to
be careful in reading the solution for Part Two. It was
minimum and maximum values and not the starting and ending
ones.