(Partially coded) Solution for Part Two - Day 16, 2020
This contains the solution for Part Two Day 16 but the code is not complete. Looks like I was
indexing wrong. Or had the index into another index but not using the correspondng fields values. Then I had
an one based index and thius grabbed the wrong values. The last comment has the correct answer. Need to come
and put that last remaining part into code.