Added code to check the expected answer against the problem solution
I really don't like the format of this. It works but too much typing.
And the checking is repeating code. There are a few things I would
change with this but going to leave for now .. would like to add a task
to the project and some utility functionality like building up a
template code.