Solution to Day 3 part 1 and part 2!
Kind of wished I had separated out the solution for part 1 from part 2
within commits. For part 2 I decided to try the generic solution for N
elves in a group instead of the specific 3 elf groups. This code could
be cleaned up some more. The seperted out of lines in the code to both
groups and individual contents of a sack is repeatative. Was trying to
do it all in one line but this may be an instance of too much cleverness
within kist comprehesion which makes the code unreadable.