Made some tweaks to how parsed through and ended the parsing
The biggest issue was finnishing off the parsing. The logic would take
itup to the last line and then get stuck there. I resolved it, can't
remember all the details, but I did add a line to the input to help
close it off.

In running this against the input file realized an issue with the
overall assumption I made which was all directories were unique. I was
trying to ignore the cd movement but a directory could have a
subdirectory with the same name. So the "quick method" was not only not
quick it was also a very bad assumption. I do need to parse as a
relative tree. I should note I don't like "tree" problems.