OK .. getting closer .. but ..
I am getting closer. Anyone following along should start to see why I
dislike "tree" problems. I struggle with how to structure and map
recursive methods and data structures. The generl concept I get. But to
move that into a this is how the structure will be and this is how the
method goes through is tought for my to work out in my head. So we shall
see several interations ..

Also just realized I've done the same thing as the previous work. I made
a flat structure of directories in a dictionary .. which means .. I will
have the same problem that a parent - child with same name directory
will be lost .. back to the drawing board ..