Changed the way I built the traversed path
I know that I would have an issue once I moved past the root directory
with the shortcut method I was using to make a traversed path string.
For example, given the made up path

```
pwp = ['/', 'a', 'b', '..', 'c', 'd'
traversed_path = '/'.join(pwp)
```

would give me an initial double forward slash. I was thinking I could
write out some iterator that wouldn't do so if the path was '/' (and
might need to in the future if the commands don't traverse from parent
to child but jump somewhere). But I wanted to simplify it. So I cam up
with and thought simply skiping the first part would work .. ah hah
That explains the two entries initially and goes back to indexing out of
range. So I did

```
traversed_path = '/'.join(pwp[1:])
```

Well, when pwp is just '/' then pwp[1:] is ''. The other one came from
the initialization of fs [line 73]. Ok that makes sense now. I am going
to leave this where it create a double slash and just handle it on the
processing end!