Initial Day 8 solution
Trying to decide and work out methods for viewing with varius form of
efficiency. For example I could alway look left and just flip or rotate
the matrix. Also if you find a unique max (that is a maxiumum value
which is not repeated) in row does one need to check the other
direction? I could see some optimization. But for now porbably go for
brute force way.