Modified to speed up operation
Using Jonathan Paulson's solution of the eval operation and his lcm
modification. I say for some reason his lcm greatly reduces the
computation time which I don't yet quite understand. I haven't watch his
whole video so maybe he is explaining there. I do like the eval
operation as well as then on does not have to deal with whether or not
an operand is old of a number. The lambda function takes care of that.

Getting wrong answer right now so thinking of going back to the list of
items per monkey.