Fixed issue with mixed type comparison
First it looks like, from the rules, that if there a mixed type which
should just be converting the int to a list and comparing just the first
item in the list/converted list. I was not initially doing this but
instead converting the singular int to a list with one item and comparing
them as lists; which appears to be incorrect. As the compare method is
expecting lists I needed to add these lines (in this commit) to extract
just the first item in the list.

But running against the puzzle input my code hits a case where though
that list is an empty list (so a number is compared to an empty list).
Thus my indexing `litem = [litem[0]] is throwing an out of range error
on `litem[0]`. I think I can resolve this by indexing a range which if
out of range should return an empty list.