Rough solution for Day 15: Part 2
This solution gives a few possibilities for answers to Part 2. It looks
for gaps in the y ranges. What is doesn't do, and I suspect this is
where the additional answers are, is to check to see if any of these
gaps are filled by beacons. This is a ToDo is solution is revisted and
cleaned up ...