Solution for Part 1 Day 3
Reworked the way I parsed out the numbers and symbols. The previous
method had a few issues. One if a symbol came right before or after a
number the split on `.` would not have separated the number from the
symbol. Also as I did this on a flattened schematic is a number was on
the end of line and one on the beginning of the next line those would
have been ccombined. Finally if two symbol were next to each other the
split would have also counted this as one symbol. So switched to a more
character by character parsing.