Added pseudo code for one method to count all the cards up
I know there are a few methods.One which seems the best would be some
recursive/generator type parse through and count. But admittedly my
create generator in Python is a bit weak. I am going to ask in a user
group about this but in the mean time I think this "algorithm" should
proper count out the total number of cards.