Searching for number of steps in a A Z matrix
Slightly different method, now looping through the A nodes and seeing
how many steps each Z node takes. But it looks like some of the a nodes
might not connect to some of the Z nodes. Guessing there is some endless
loops .. which I think one could detect. Keep track of which nodes you
went on and which direction you took on that node. If you see a repeat
before you get the the end node then you should be in a loop.