Added check to see if a loop was discovered
The idea was that if we are on a node we have visted before then we must
be in a loop. This code appeared to worke for the test input although I
need to double check that. But for the problem input found loops on all
paths .. do I need to check the index of the L/R instruction instead ..
??