Added fix for seeing start as a neighboring position
Added in code to handle the starting position. Not sure if I really like
or dislike this solution but will work with it as is till I need to
adjust.

Did discover I am skipping the corners and moving ahead to the next
step which I see now. I can suimplify the code to fix this although I
like the way this reads now (even though it is faulty).