.. and this is getting to be no fun.
Got a solution .. yes it is pretty much brut force .. and it seems
pretty sound. Trying to pull out any tasks that I could once up front
and not repeat it a billion times. But this is just getting no fun. And
then once I complete it I need to verify that it will not crash with a
billion cycles. So I am moving on .. will comback to this .. maybe ..