Initial work on Day 16
Thinking through the problem some more. I have initially believed I want
each of the reflectors (mirrors and splitters) to recall from where a
beam has come from. That way I can easily see if a beam has already come
in from that direction. That is it looping? If so we can kill off that
beam.

Initially I was going to have a collection of beams. Each beam was going
to a list of indexes to visit. Then is was going to be a starting
position and a direction. The direction was going to help when it
encountered a reflector; we could easily translate into the reflected
direction or action.

At this point I realized that other than the initial start all beams
would emit from a reflector. So I am thinking I could use the reflector
structure to also keep track of the beams.