Fleshed out and coded a bit more design
In writing out the initial path to the first reflector I saw this would
be the same from any path. So made this a function, `find_next_reflector`,
which traces out a path and depending on the direction which reflectors
to ignore. (That is the pointy ends of splitters). Then it looks down
that path to find the next reflector. If no reflector is found it will
return None as the next reflectore. It also returns the tiles it
traversed.

Next we will need to react to a beam hitting a reflector ..