Added initial reflection method
Coded out the initial method to handle the reflection of an incoming
beam onto the next reflector. This starts to raise the question o how
and when or where do we remove a beam. (Thinking this may be on the find
next reflector method ..)

Also cleaned up the static variables for the strings.