A few more changes ..
- Added code to remove beams
- Added note (and will eventually add code) to skip reflecting a beam if
  one has arrive ath this reflector from the same direction as a method
  to prevent infinite looping
- Updated code around initial shinning of light; filling in arguments
  and passing along data to the reflect method

As I work this problem I see that my solution, for this problem, might
be best as a class based solution. There are a few reasons for this, in
particular the reflectors object is shared and there is other shared
variables and data. Going to shift the code and try this out ..