Switched method for recording new beams
Instead of popping outgoing beams of the current list as one looks for
next reflector and then readding on reflection, I am going to record
what needs to be added and then add in after all reflectors reviewed.
The current implementation within this commit has an issue in that right
now it is attibuting the new beams to the urrent reflector and not, as
it should, the next reflector. As we look at the new beams through the
lense of the current refelctors beaming out onto the next reflector I
will need to gather them all up and the post process them.