Fixed issue with update beams algorithm
Resolved the previous issues with the update algorithm. I have a few
extra methods here in the class and need to do some general cleaning up
but this is running again. I am getting a similar answer as before which
is 22 tiles energized. Need to add some more debugging. Might add a
similar map at see what the end result is visualized.