Solution for Part 1 Day 16
Initially when running against the problem data I was getting too high a
count. Looking at debugging output and finally looking at the problem
data set I see that the initial start of the data reflects on tile zero
which my initial start did not handle. Just changing the starting
position to -1 did not fully resolve as then the initial path as was
calculated would have been empty thus ending the reflections. So added a
small hack that I might need to leave in there. I of course could copy
out the look right and find next reflector as an initial find reflector,
removing the check to see if pos is -1. But this one I may leave. I want
to clean up this solution to get it back to the pretty solid state it
was in.