Readded check to see if any reflector found on inital path
With being able to start in any edge piece it is now possible again to
not encounter a reflector in the inital path. Readded in code to handle
this case.

Now have an issue with what happens if on on bounce. Seems I am missing
a scenario and trying an empty set of beams ..