Resolved issue that broke part 1 solution
Found it was my logic on the initial find next reflector as to whether
or not reflect on that forst one. As the first reflector in the problem
set is at position zero checking for if relector is the same as if 0
which fails and is not what I want. Instead I corrected by checking to
see if next_reflector is not None.