Day 1 Part 1 solution ..
Ok, just used modulus operator. Initially thought of this but tried soemmthing real quick and it seemed
negatives numbers mod 100 was different than positive numbers mod 100. So went initial with the let's cheat
with the index wrapping on negative numbers. But that fails outside of index range +/- 100 .. so back to modulus.
And It seems I just did something silly to think I needed to handle negative numbers as well .. oh well