Adjusted so one was not applying adjustment twice
There is still an issue if the pointer is zero and it goes negative. In this case we don't want to pick up another number of zeros.
Also I suspect an issue if the pointer is negative and the adjustment is zero. Might be able to compensate by if adjustment is zero just proceed to the next.
Not sure about on the positive side ..