Solution in progress ..
Working towards a solution for part 2 of day 2. Seeing I could make this a bit
more generic. I might do that or just continue done this path.

Also seeing an issue with what I have when the factor is two. So going to save this
as is and look into it.