Near to solution ..
Have it there a bit more. Realized using the test data that I forgot to handle situations like
`222222` which will match on factors of 1, 2, and 4. but only want to count. Think the quick
solution is to just set them to make the list unique.

Also noticing an is with just double digits .. fixing now.