So ..
this type of problem (even part 1) is the tyope I am simply not interested
in. A bit of my mindset .. This is and primarily is a game. Yes one gets
some programming practice oout of it. But still the end goal is to solve a
problem. Problem is not all problems interest me. So not ever one is one I
want to spend doing (especially for a game - a side activity to many of life's
more interesting things to do ..). And this happens to be one of those I am
not really interested in. Another day might circle back on this ..